Saving latest Updates with script

+++ home tutor/app/src/main/java/com/example/login_page1/Extensions.kt
@@ -80,7 +80,7 @@ fun isValidTutorAge(age: Int): Boolean {
     return age in MIN_TUTOR_AGE..MAX_TUTOR_AGE // Assuming a reasonable age range
 }
 
-fun isValidPhNo(phNo : Int) : Boolean{
+fun isValidPhNo(phNo : Long) : Boolean{
     //Context.pr("PhNo chk")
     val regex = Regex(PHONE_REGEX)
    // return true
